Day 3 Final Commit I guess
I had to turn it in. I was not able to get the words to appear as you were typing them up. I did do the site differently then It asked us to do it because I thought it would be better for kinder-gardeners. I also changed the solved button to a reset button