Final Commit
I got the word typing thing to work. I did do it differently the the book wanted but it is better I think for younger kids and it discourages cheating by making it hard to go through and click every single key. Thank you for the extra day for work.